# Conflicts:
#	README.md
#	server/core/core/src/de/uniol/inf/is/odysseus/core/server/OdysseusConfiguration.java
#	server/core/core/src/de/uniol/inf/is/odysseus/core/server/logicaloperator/CoalesceAO.java
#	server/crypt/de.uniol.inf.is.odysseus.incubation.crypt.keymanagement.server/src/de/uniol/inf/is/odysseus/incubation/crypt/keymanagement/server/KeyWebSocketServer.java
#	server/keyvalue/keyvalue/src/de/uniol/inf/is/odysseus/server/keyvalue/physicaloperator/KeyValueSelectPO.java


Former-commit-id: 08ee199754449343f189d81cdc5c8ad373c0a04b

+++ wrapper/telegram/src/de/uniol/inf/is/odysseus/wrapper/telegram/TelegramTransportHandler.java
@@ -112,7 +112,7 @@ public class TelegramTransportHandler extends AbstractPushTransportHandler {
 			@SuppressWarnings("unchecked")
 			KeyValueObject<IMetaAttribute> kvMessage = (KeyValueObject<IMetaAttribute>)message;
 			// TODO: Why the cast?
-			Long chatID=  kvMessage.getAttribute("chat.id");
+			Integer chatID=  kvMessage.getAttribute("chat.id");
 			if (chatID == null){
 				throw new RuntimeException("Cannot find attribute with name chat.id");
 			}
